refactor: split file products.ts into modules
