+++ src/index.ts
@@ -1,6 +1,4 @@
 import { start } from './pages/products/products';
 import './style.css';
 
-console.log('Online-Store');
-// getList().then(getProductsList);
 start();
